service classes for seperate user registrations
