+++ src/Pages/Authentication/Register_Forms/Restaurant_Register.jsx
@@ -1,27 +1,22 @@
 import { useRef, useState, useEffect } from "react";
 import { faCheck, faTimes, faInfoCircle } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-//import Login from './Login';
 import { useNavigate } from "react-router-dom";
-import loginService from "../../Services/loginService";
-
+import loginService from "../../../Services/loginService";
 
 const PHONE_REGEX = /^[0][0-9]{9}$/;
 
-const Customer_Register = () => {
-
+const Restaurant_Register = () => {
     const phoneRef = useRef();
     const errRef = useRef();
 
     const [name, setName] = useState('');
-
+    const [location, setLocation] = useState('');
+    const [contact, setContact] = useState('');
     const [email, setEmail] = useState('');
 
-    const [phone, setPhone] = useState('');
     const [validPhone, setValidPhone] = useState(false);
     const [phoneFocus, setPhoneFocus] = useState(false);
-
-    const [address, setAddress] = useState('');
 
     const [errMsg, setErrMsg] = useState('');
     const [success, setSuccess] = useState(false);
@@ -31,64 +26,54 @@ const Customer_Register = () => {
     }, [])
 
     useEffect(() => {
-        setValidPhone(PHONE_REGEX.test(phone));
-    }, [phone])
-
+        setValidPhone(PHONE_REGEX.test(contact));
+    }, [contact])
 
     useEffect(() => {
         setErrMsg('');
-    }, [name, email, phone, address])
+    }, [name, location, contact, email])
 
     const handleSubmit = async (e) => {
         e.preventDefault();
         // if button enabled with JS hack
-        const v1 = PHONE_REGEX.test(phone);
-        if (!v1) {
+        const validContact = PHONE_REGEX.test(contact);
+        if (!validContact) {
             setErrMsg("Invalid Entry");
             return;
         }
         try {
-            // Make an API call to create a user using loginService
-            const response = await loginService.createUser({
-                name: name,
-                email: email,
-                phone: phone,
-                address: address,
+            // Make an API call to create a restaurant using loginService
+            const response = await loginService.createRestaurant({
+                name,
+                location,
+                contact,
+                email,
             });
-    
+
             // Log the response data and access token
             console.log(response?.data);
             console.log(response?.accessToken);
             console.log(JSON.stringify(response));
-            
-    
+
             // Set success state and clear input fields
             setSuccess(true);
             setName('');
+            setLocation('');
+            setContact('');
             setEmail('');
-            setPhone('');
-            setAddress('');
-    
-            
-          
-            
-            
 
         } catch (err) {
-            
             if (!err?.response) {
                 setErrMsg('No Server Response');
             } else if (err.response?.status === 409) {
-                setErrMsg('Phone number Taken');
+                setErrMsg('Contact number Taken');
             } else {
-                setErrMsg('Registration Failed')
+                setErrMsg('Registration Failed');
             }
-            
+
             errRef.current.focus();
         }
     }
-
-    
 
     return (
         <div style={{ display: "flex", justifyContent: "center", alignItems: "center", height: "100vh" }}>
@@ -102,77 +87,57 @@ const Customer_Register = () => {
             ) : (
                 <section>
                     <p ref={errRef} className={errMsg ? "errmsg" : "offscreen"} aria-live="assertive">{errMsg}</p>
-                    <h1>Customer Register</h1>
+                    <h1>Restaurant Register</h1>
                     <form onSubmit={handleSubmit}>
-                        <label htmlFor="name">
-                            Name:
-                        </label>
+                        <label htmlFor="name">Name:</label>
                         <input
                             type="text"
                             id="name"
                             autoComplete="off"
                             onChange={(e) => setName(e.target.value)}
-                            value={name}aria-describedby="namenote"
+                            value={name}
                             required
-                            
                         />
-                        
 
-                        <label htmlFor="email">
-                            Email:
-                            
-                        </label>
+                        <label htmlFor="location">Location:</label>
                         <input
                             type="text"
-                            id="email"
-                           // ref={emailRef}
+                            id="location"
                             autoComplete="off"
-                            onChange={(e) => setEmail(e.target.value)}
-                            value={email}required
+                            onChange={(e) => setLocation(e.target.value)}
+                            value={location}
+                            required
                         />
 
-                   
-
-
-                        <label htmlFor="phone">
-                            Phone:
-                            <FontAwesomeIcon icon={faCheck} className={validPhone ? "valid" : "hide"} />
-                            <FontAwesomeIcon icon={faTimes} className={validPhone || !phone ? "hide" : "invalid"} />
-                        </label>
+                        <label htmlFor="contact">Contact:</label>
                         <input
                             type="text"
-                            id="phone"
-                            onChange={(e) => setPhone(e.target.value)}
-                            value={phone}
-                            
+                            id="contact"
+                            onChange={(e) => setContact(e.target.value)}
+                            value={contact}
                             ref={phoneRef}
                             required
-                            aria-invalid={validPhone ? "false" : "true"}
-                            aria-describedby="phonenote"
+                            aria-invalid={!validPhone}
+                            aria-describedby="contactnote"
                             onFocus={() => setPhoneFocus(true)}
                             onBlur={() => setPhoneFocus(false)}
                         />
-                        <p id="phonenote" className={phoneFocus && !validPhone ? "instructions" : "offscreen"}>
+                        <p id="contactnote" className={phoneFocus && !validPhone ? "instructions" : "offscreen"}>
                             <FontAwesomeIcon icon={faInfoCircle} />
                             Start with 0.<br />
                             Must include 10 numbers.<br />
                         </p>
 
-
-                        <label htmlFor="address">
-                            Address:
-                        </label>
+                        <label htmlFor="email">Email:</label>
                         <input
                             type="text"
-                            id="address"
-                            onChange={(e) => setAddress(e.target.value)}
-                            value={address}
+                            id="email"
+                            onChange={(e) => setEmail(e.target.value)}
+                            value={email}
                             required
                         />
-                        
 
-                        
-                        <button disabled={!validPhone? true : false}>Sign Up</button>
+                        <button disabled={!validPhone ? true : false}>Sign Up</button>
                     </form>
                     <p>
                         Already registered?<br />
@@ -186,6 +151,4 @@ const Customer_Register = () => {
     )
 }
 
-export default Customer_Register
-
-
+export default Restaurant_Register;
