+++ src/Pages/Authentication/Register_Forms/ShopOwner_Register.jsx
@@ -2,17 +2,18 @@ import { useRef, useState, useEffect } from "react";
 import { faCheck, faTimes, faInfoCircle } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { useNavigate } from "react-router-dom";
-import loginService from "../../Services/loginService";
+import loginService from "../../../Services/loginService";
 
 const PHONE_REGEX = /^[0][0-9]{9}$/;
 
-const Restaurant_Register = () => {
+const ShopOwner_Register = () => {
     const phoneRef = useRef();
     const errRef = useRef();
 
-    const [name, setName] = useState('');
-    const [location, setLocation] = useState('');
+    const [shopName, setShopName] = useState('');
     const [contact, setContact] = useState('');
+    const [branch, setBranch] = useState('');
+    const [location, setLocation] = useState('');
     const [email, setEmail] = useState('');
 
     const [validPhone, setValidPhone] = useState(false);
@@ -31,7 +32,7 @@ const Restaurant_Register = () => {
 
     useEffect(() => {
         setErrMsg('');
-    }, [name, location, contact, email])
+    }, [shopName, contact, branch, location, email])
 
     const handleSubmit = async (e) => {
         e.preventDefault();
@@ -42,11 +43,12 @@ const Restaurant_Register = () => {
             return;
         }
         try {
-            // Make an API call to create a restaurant using loginService
-            const response = await loginService.createRestaurant({
-                name,
-                location,
+            // Make an API call to create a shop owner using loginService
+            const response = await loginService.createShopOwner({
+                shopName,
                 contact,
+                branch,
+                location,
                 email,
             });
 
@@ -57,9 +59,10 @@ const Restaurant_Register = () => {
 
             // Set success state and clear input fields
             setSuccess(true);
-            setName('');
-            setLocation('');
+            setShopName('');
             setContact('');
+            setBranch('');
+            setLocation('');
             setEmail('');
 
         } catch (err) {
@@ -87,25 +90,15 @@ const Restaurant_Register = () => {
             ) : (
                 <section>
                     <p ref={errRef} className={errMsg ? "errmsg" : "offscreen"} aria-live="assertive">{errMsg}</p>
-                    <h1>Restaurant Register</h1>
+                    <h1>Shop Owner Register</h1>
                     <form onSubmit={handleSubmit}>
-                        <label htmlFor="name">Name:</label>
+                        <label htmlFor="shopName">Shop Name:</label>
                         <input
                             type="text"
-                            id="name"
+                            id="shopName"
                             autoComplete="off"
-                            onChange={(e) => setName(e.target.value)}
-                            value={name}
-                            required
-                        />
-
-                        <label htmlFor="location">Location:</label>
-                        <input
-                            type="text"
-                            id="location"
-                            autoComplete="off"
-                            onChange={(e) => setLocation(e.target.value)}
-                            value={location}
+                            onChange={(e) => setShopName(e.target.value)}
+                            value={shopName}
                             required
                         />
 
@@ -127,6 +120,26 @@ const Restaurant_Register = () => {
                             Start with 0.<br />
                             Must include 10 numbers.<br />
                         </p>
+
+                        <label htmlFor="branch">Branch:</label>
+                        <input
+                            type="text"
+                            id="branch"
+                            autoComplete="off"
+                            onChange={(e) => setBranch(e.target.value)}
+                            value={branch}
+                            required
+                        />
+
+                        <label htmlFor="location">Location:</label>
+                        <input
+                            type="text"
+                            id="location"
+                            autoComplete="off"
+                            onChange={(e) => setLocation(e.target.value)}
+                            value={location}
+                            required
+                        />
 
                         <label htmlFor="email">Email:</label>
                         <input
@@ -151,4 +164,4 @@ const Restaurant_Register = () => {
     )
 }
 
-export default Restaurant_Register;
+export default ShopOwner_Register;
